Breaking: make distribution its own construct and overrideable
